% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_tol.R
\name{define_tol}
\alias{define_tol}
\title{Define the tolerance value to be used in ABC model}
\usage{
define_tol(
  comm.obs.meta,
  phylo.meta,
  envir,
  xy.coords,
  Nmeta,
  parallel,
  m = 0.5,
  W.r.prior = FALSE,
  OU.alpha = "uniform",
  tol = 1,
  n.timestep = 50,
  max.sample.size.prior = 100 * parallel,
  max.sample.size.posterior = 20 * parallel,
  scenario.ID = "mcfly",
  output.dir.path = "meta_output"
)
}
\arguments{
\item{comm.obs.meta}{Matrix containing the observed metacommunity.}

\item{phylo.meta}{Newick object with phylogenetic relationship among species of metacommunity.}

\item{envir}{Vector with values of environmental variable describing each community of metacommunity.}

\item{xy.coords}{Two column matrix with coordinates of communities in metacommunity.}

\item{Nmeta}{Numeric indicating the number of metacommunities to be simulated in ABC model.}

\item{parallel}{Numeric value indicating the number of cores to be used in parallel computation.}

\item{m}{Numeric value indicating the immigration rate at each site, reported as Hubbel´s m. This is the same parameter accepted by \code{\link{metasim}}.}

\item{W.r.prior}{Logical (TRUE or FALSE) indicating if the the W.r parameter would be a single value (FALSE) with value of 0, indicating a panmictic metacommunity
or follow a prior distribution (TRUE) of values calculated as being the slopes of dispersal kernel indicating the contribution of species from neighboring patches
to the local immigrant pool.}

\item{OU.alpha}{Character indicating the type of prior that will be used in ABC model. The options were "uniform" for a uniform sample of
alpha values and "half-life" for a prior of alpha values represented as being half-life values, calculated as being log().}

\item{tol}{Numeric value indicating the initial tolerance to be used in mcfly function. Default is 1}

\item{n.timestep}{Numeric value indicating the number of timesteps to simulate the metacommunity.
same argument used in \code{\link{metasim}} function.}

\item{scenario.ID}{Character indicating the name of the simulation scenario. The same as used in \code{\link{metasim}}. Default is "mcfly".}

\item{output.dir.path}{Character indicating the name of directory to save simulations results and metadata used in \code{\link{metasim}}. Default is "delorean".}

\item{max.sample.prior}{Numeric value indicating the maximmum size of sampled prior distribution. Default is 100*parallel.}

\item{max.size.posterior}{Numeric value indicating the maximum size of posterior distribution. Default is 20*parallel.}
}
\value{
A matrix containing the total sample size used to find the posterior distribution, the length of posterior distribution,
the minimum alpha value that can be sample given the phylogenetic tree used,
the maximum alpha value that can be sample given the phylogenetic tree used,
the highest 1\% of tolerance values found in ABC model and the alpha values sampled from
prior distribution.
}
\description{
Auxiliary function to define tolerance value to be used in the mcfly function.
}
\details{
This function provides a procedure to estimate the values of tolerance that
can be used in \code{\link[=mcfly]{mcfly()}} function, since the tolerance value will vary accordingly to
different alpha values.
}
\examples{
\dontrun{
    
     # read datasets 
     comm <- data("Furnariidae") # community matrix
     phylo <- data("phylo_Furnariidae") # phylogenetic hypothesis
     envir_furnariidae <- data("envir") 
     coords <- envir_furnariidae[, c(1, 2)] # site coordinates
     envir <- envir_furnariidae[, -c(1, 2)] # environmental variables
     # run define_tol function
     res_tol <- define_tol(comm.obs.meta = comm,
           phylo.meta = phylo, 
           envir = envir, 
           xy.coords = coords, 
           parallel = NULL
           )
     tol_val <- res_tol[, 5]
           
    }
}
\seealso{
\code{\link{mcfly}}
}
