% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dimensionality_internal_mcfly.R
\name{dimensionality}
\alias{dimensionality}
\title{dimensionality}
\usage{
dimensionality(
  matrix.M,
  scale = TRUE,
  method = "standardize",
  evenness = "Camargo"
)
}
\arguments{
\item{matrix.M}{Matrix object containing diversity metric values for each community.}

\item{scale}{Logical argument. If TRUE (default) the diversity metrics in matrix M are standardized.}

\item{method}{Character. The method of standardization to be used in matrix M.}

\item{evenness}{Character. Evennes metric that will be used to compute EE. Default is "Camargo" for Camargo index.}
}
\value{
Numerical scalar with value of EE.
}
\description{
dimensionality
}
