% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{envir}
\alias{envir}
\title{Environmental variable}
\format{
matrix
}
\source{
Maestri and Duarte 2020 \url{https://link.springer.com/chapter/10.1007/978-3-030-31167-4_6}
}
\description{
Matrix containing environmental descriptor for the sites.
}
