% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Furnaridae}
\alias{Furnaridae}
\title{Furnaridae data set}
\format{
A data frame with 1350 rows and 102 columns:
}
\source{
Maestri and Duarte 2020 \url{https://link.springer.com/chapter/10.1007/978-3-030-31167-4_6}
}
\usage{
Furnaridae
}
\description{
Data frame containing occurence of 102 species of Furnaridae in 1350 sample units
}
\keyword{datasets}
