% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ImportanceVal_V2_internal_mcfly.R
\name{ImportanceVal}
\alias{ImportanceVal}
\title{Importance Values}
\usage{
ImportanceVal(matrix.M, scale = TRUE, method = "max", stopRule = TRUE)
}
\arguments{
\item{matrix.M}{Matrix object containing diversity metric values for each community.}

\item{scale}{Logical argument. If TRUE (default) the diversity metrics in matrix M are standardized.}

\item{method}{Character. The method of standardization to be used in matrix M.}

\item{stopRule}{Logical. If TRUE a criterion for select eigenvalues will be used.}
}
\description{
Importance Values
}
