% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcfly.R, R/mcfly_RUN_internal.R
\encoding{UTF-8}
\name{mcfly}
\alias{mcfly}
\alias{f.internal}
\title{mcfly}
\usage{
mcfly(
  comm,
  phylo,
  envir,
  xy.coords,
  occurrence = TRUE,
  entropy.order = 1,
  niche.breadth = 10,
  m = 0.5,
  n.timestep = 50,
  OU.alpha = c("uniform", "half-life"),
  W.r.prior = FALSE,
  summary.stat = "correlation",
  tol = 0.2,
  sample.size.posterior = 240,
  max.sample.size.prior = 2400,
  HPD = 0.9,
  return.comm = FALSE,
  return.w.priors = FALSE,
  return.alpha.priors = FALSE,
  parallel = NULL,
  scenario.ID = "mcfly",
  output.dir.path = "delorean"
)

f.internal(
  k,
  sample.size.posterior,
  max.sample.size.prior,
  prior.alpha,
  prior.w,
  theta.val,
  phylo,
  niche.sigma,
  root.value,
  my.landscape,
  JM,
  n.timestep,
  spp.freq,
  niche.breadth,
  occurrence,
  spp.names,
  names.comm,
  entropy.order,
  summary.stat,
  div,
  tol,
  return.comm,
  return.w.priors,
  return.alpha.priors,
  scenario.ID,
  output.dir.path
)
}
\arguments{
\item{comm}{Matrix containing occurrences or abundances of species in sites. Species in columns and sites in rows.}

\item{phylo}{phylogenetic tree in newick format}

\item{envir}{A one column matrix containing environmental variable for each community}

\item{xy.coords}{A two column matrix containing the coordinates of each community}

\item{occurrence}{if occurence or abundance}

\item{entropy.order}{entropy order for Rényi series}

\item{niche.breadth}{niche breadth}

\item{m}{Numeric value indicating the immigration rate at each site, reported as Hubbel´s m. This is the same parameter accepted by \code{\link{metasim}}.}

\item{n.timestep}{number of timesteps for metacommunity simulation}

\item{OU.alpha}{Character indicating the type of prior that will be used in ABC model. The options were "uniform" for a uniform sample of
alpha values and "half-life" for a prior of alpha values represented as being half-life values, calculated as being log().}

\item{W.r.prior}{Logical (TRUE or FALSE) indicating if the the W.r parameter would be a single value (FALSE) with value of 0, indicating a panmictic metacommunity
or follow a prior distribution (TRUE) of values calculated as being the slopes of dispersal kernel indicating the contribution of species from neighboring patches
to the local immigrant pool.
Same as accepted by \code{\link{metasim}}.}

\item{summary.stat}{type of summary statistic}

\item{tol}{tolerance value}

\item{sample.size.posterior}{size of posterior distribution}

\item{max.sample.size.prior}{maximum sample size for prior distribution}

\item{HPD}{Numeric value indicating the probability mass for the Highest Density Interval for the posterior
probability distribution obtained in ACB model. This is the same value used in \code{\link{hdi}}. Default is 0.9.}

\item{return.comm}{logical}

\item{return.w.priors}{Logical (TRUE/FALSE), indicating if the prior distribution of W.r values used in ABC model must be returned in the output.
Default is FALSE}

\item{return.alpha.priors}{Logical (TRUE/FALSE), indicating if the the prior distribution of alpha values must be returned in the output. Default is FALSE.}

\item{parallel}{Numerical value indicating the numbers of cores that must be used in the parallel computation. Default is NULL, indicating that the
calculations of ABC model will not be parallelized.}

\item{scenario.ID}{name of simulation scenario}

\item{output.dir.path}{name of output directory}

\item{k}{only for parallel computation}

\item{prior.alpha}{type of prior alpha, the same as in mcfly}

\item{prior.w}{type of prior w}

\item{theta.val}{theta}

\item{niche.sigma}{sigma}

\item{root.value}{root value}

\item{my.landscape}{simulated landscape}

\item{JM}{JM parameter, the same accepted by \code{\link{metasim}}}

\item{spp.freq}{species frequency}

\item{spp.names}{species names}

\item{names.comm}{communities names}

\item{div}{diversity}
}
\value{
List, with the following components:
\item{Time.spent}{Matrix with initial and final date and time spend to estimate parameters in ABC model.}
\item{COMM.sim}{Simulated communities, returned only if if return.comm is TRUE.}
\item{Species.Pools}{Matrix containing the number of species in phylogeny and in metacommunity}
\item{Sample_Attributes}{Matrix containing the attributes of simulation. Size of priors and size of posterior distribution}
\item{Alpha_Limits}{Matrix containing minimum the root phylogeny age, minimum and maximum alpha values sampled in the uniform distribution.}
\item{Alpha.prior.mode}{Numeric vector with alpha values used in the prior distribution. Represented as alpha (default) or
half-life values of alpha (if argument OU.alpha is set as being "half-life")}
\item{W_Prior_Distribution}{Numeric. A scalar with the value of W.r parameter (default), or a vector if argument W.r.prior is TRUE.}
\item{Theta}{Numeric indicating the value of theta parameter used in ABC model.}
\item{K_niche}{Numerical indicating the phylogenetic signal in niche position estimated accordingly to K statistic. The same value returned in
\code{\link{Kcalc}} function.}
\item{Summary.Statistics}{Numerical vector with values of correlation statistic between diversity values of simulated and empirical metacommunities.}
\item{Alpha_Posterior_Distribution}{Numerical vector with alpha values obtained from posterior distribution.}
\item{HPD_Alpha}{Numerical vector of lenght two with lower and upper limits of the HPD distribution for alpha parameter.}
\item{W_Posterior_Distribution}{Numerical vector with W.r values of posterior distribution.}
\item{HPD_w}{Numerical vector of length two with lower and upper limits of the HOD distribution for W.r parameter.}
}
\description{
mcfly function to estimate the influence of stabilizing niche selection on species diversity across environmental gradients
}
\details{
This function estimate the influence of stabilizing niche selection on species diversity across environmental gradients by applying to
a occurrence matrix of species containing presence/absence or abundance an Approximate Bayesian Computation (ABC) framework. We used in ABC a individual
based-model from MCSim package
}
