% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_list_function.R
\name{organize.list}
\alias{organize.list}
\title{Internal function}
\usage{
organize.list(x)
}
\arguments{
\item{x}{List object returned from \code{\link{parallel_PAR}}}
}
\value{
Matrix with results obtained from simulation proccess
}
\description{
Auxiliar function to organize the results from parallel computation
}
