% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAR_function.R
\name{parallel_PAR}
\alias{parallel_PAR}
\title{parallel_PAR}
\usage{
parallel_PAR(
  OU.alpha.v,
  comm,
  subset,
  occurrence,
  tree,
  mat.env,
  landscape,
  JM,
  n.timestep,
  spp.freq,
  W.r,
  sigma,
  theta,
  root.value,
  scenario.ID,
  sim.ID,
  output.dir.path,
  reps = 5
)
}
\arguments{
\item{OU.alpha.v}{Numeric vector containing values of alpha parameter to be tested}

\item{comm}{Matrix with species occurence or abundances, sampling units in rows and species in columns}

\item{subset}{Logical, if TRUE only a subset of species will be used in caculations}

\item{occurrence}{Logical, if TRUE community matrix will be converted to species occurence}

\item{tree}{Phylogenetic tree in newick format}

\item{mat.env}{Matrix containing values for environmental gradient}

\item{landscape}{Landscape where the simulations will go on. Object generated with \code{\link{make.landscape}} from MCSim package}

\item{JM}{Numeric value indicating the total number of individuals to include in \code{\link{make.landscape}} from MCSim package}

\item{n.timestep}{Numeric value indicating the number of generations in the simulation process performed with \code{\link{metasim}} from MCSim package}

\item{spp.freq}{Numeric vector of regional abundance, used in \code{\link{metasim}} from MCSim package}

\item{W.r}{Numeric indicating the dispersal kernel slope in \code{\link{metasim}} from MCSim package}

\item{sigma}{Numeric vector indicating the standard-dviation of niche position. Argument used in \code{\link{rTraitCont}}}

\item{theta}{Numeric value used in  \code{\link{rTraitCont}} to indicate the optimun for each branch}

\item{root.value}{Numeric value indicating the trait value at the root to be used in \code{\link{rTraitCont}}}

\item{scenario.ID}{Character setting the name for the simulation scenario. Used in \code{\link{metasim}}}

\item{sim.ID}{Character string indicating the name for a particular simulation. Used in \code{\link{rTraitCont}}}

\item{output.dir.path}{Character string indicating the name of directory to save simulation results. Used in \code{\link{rTraitCont}}}

\item{reps}{tricky argument to work in parallel process}
}
\value{
A list with parameters from simulated communities
}
\description{
Auxiliar function to allow parallel computations
}
