% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylo_Furnaridae}
\alias{phylo_Furnaridae}
\title{Phylogeny of Furnaridae species}
\format{
Newick containing phylogenetic relationships of 102 species of Furnaridae family
}
\source{
BirTree \url{https://birdtree.org/}
}
\usage{
phylo_Furnaridae
}
\description{
Newick object containing the phylogenetic hypothesis of 102 species of Furnaridae
}
\keyword{datasets}
