% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{phylo_Furnariidae}
\alias{phylo_Furnariidae}
\title{Phylogeny of Furnariidae species}
\format{
Newick containing phylogenetic relationships of 102 species of Furnariidae family.
}
\source{
BirTree \url{https://birdtree.org/}
}
\description{
Newick object containing the phylogenetic hypothesis of 102 species of Furnariidae.
}
