% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_define_tol.R
\name{define_tolerance}
\alias{define_tolerance}
\title{Define the tolerance value to be used in ABC model}
\usage{
define_tolerance(
  comm,
  phylo,
  envir,
  xy.coords,
  occurrence,
  entropy.order,
  niche.breadth = 10,
  m = 0.5,
  n.timestep = 50,
  summary.stat = "correlation",
  W.r.prior,
  OU.alpha = c("uniform", "half-life"),
  max.sample.size.prior = 100 * parallel,
  sample.size.posterior = 20 * parallel,
  parallel = NULL,
  scenario.ID = "doc",
  output.dir.path = "einstein",
  probs = 0.99
)
}
\arguments{
\item{comm}{community matrix}

\item{phylo}{phylo object}

\item{envir}{Vector with values of environmental variable describing each community of metacommunity.}

\item{xy.coords}{Two column matrix with coordinates of communities in metacommunity.}

\item{occurrence}{logical}

\item{entropy.order}{numeric}

\item{niche.breadth}{numeric}

\item{m}{Numeric value indicating the immigration rate at each site, reported as Hubbel´s m. This is the same parameter accepted by \code{\link{metasim}}.}

\item{n.timestep}{Numeric value indicating the number of timesteps to simulate the metacommunity.
same argument used in \code{\link{metasim}} function.}

\item{summary.stat}{character}

\item{W.r.prior}{Logical (TRUE or FALSE) indicating if the the W.r parameter would be a single value (FALSE) with value of 0, indicating a panmictic metacommunity
or follow a prior distribution (TRUE) of values calculated as being the slopes of dispersal kernel indicating the contribution of species from neighboring patches
to the local immigrant pool.}

\item{OU.alpha}{Character indicating the type of prior that will be used in ABC model. The options were "uniform" for a uniform sample of
alpha values and "half-life" for a prior of alpha values represented as being half-life values, calculated as being log().}

\item{max.sample.size.prior}{numeric}

\item{sample.size.posterior}{numeric}

\item{parallel}{Numeric value indicating the number of cores to be used in parallel computation.}

\item{scenario.ID}{Character indicating the name of the simulation scenario. The same as used in \code{\link{metasim}}. Default is "mcfly".}

\item{output.dir.path}{Character indicating the name of directory to save simulations results and metadata used in \code{\link{metasim}}. Default is "delorean".}

\item{probs}{numeric}
}
\value{
numeric indicating tolerance value
}
\description{
Auxiliary function to define tolerance value to be used in the mcfly function.
}
\details{
This function provides a procedure to estimate the values of tolerance that
can be used in \code{\link[=mcfly]{mcfly()}} function, since the tolerance value will vary accordingly to
different alpha values.
}
