% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functionParallelProcess.R
\name{f.internal}
\alias{f.internal}
\title{Internal auxiliar function to parallel process in mcfly}
\usage{
f.internal(
  k,
  sample.size.posterior,
  max.sample.size.prior,
  prior.alpha,
  prior.w,
  theta.val,
  phylo,
  niche.sigma,
  root.value,
  my.landscape,
  JM,
  n.timestep,
  spp.freq,
  niche.breadth,
  occurrence,
  spp.names,
  names.comm,
  entropy.order,
  div,
  tol,
  return.comm,
  scenario.ID,
  output.dir.path
)
}
\arguments{
\item{k}{numeric, auxiliar parameter to paralell proccess}

\item{sample.size.posterior}{numeric indicating the size of posterior distribution}

\item{max.sample.size.prior}{numeric indicating the sie of prior distribution}

\item{prior.alpha}{character indication the type of prior alpha same as mcfly}

\item{prior.w}{logical indicating if the w must be estimated}

\item{theta.val}{numeric theta value}

\item{phylo}{phylogeny}

\item{niche.sigma}{numeric indicating the sigma for nich}

\item{root.value}{numeric indicating the mean value of traits}

\item{my.landscape}{landscape}

\item{JM}{total number of individuals in simulation process}

\item{n.timestep}{number of timesteps in simulation process}

\item{spp.freq}{frequency of species}

\item{niche.breadth}{numeric indicating the niche breath of species}

\item{occurrence}{logical}

\item{spp.names}{names of species}

\item{names.comm}{names for communities}

\item{entropy.order}{diversity value accordingly to entropy order}

\item{div}{diversity value}

\item{tol}{tolerance value}

\item{return.comm}{logical}

\item{scenario.ID}{character}

\item{output.dir.path}{character}
}
\value{
list with results to be used in mcfly function
}
\description{
Internal auxiliar function to parallel process in mcfly
}
