% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mcfly.R, R/mcfly_RUN.R
\encoding{UTF-8}
\name{Mcfly}
\alias{Mcfly}
\alias{mcfly_RUN}
\title{Mcfly}
\usage{
Mcfly(
  comm,
  subset,
  occurrence = FALSE,
  env,
  site.coords,
  tree,
  OU.alpha,
  sigma,
  theta,
  root.value,
  runs,
  ncores = NULL,
  area.m2 = 1,
  m = 0.5,
  JM = sum(comm),
  JM.limit = JM,
  JL = rowSums(comm),
  nu = 0,
  speciation.limit = 0,
  n.timestep = 50,
  W.r = 0,
  scenario.ID = "species.sorting",
  sim.ID = "data",
  output.dir.path = "OUTPUT_DATA"
)

mcfly_RUN(
  OU.alpha.v,
  comm,
  subset,
  occurrence,
  tree,
  mat.env,
  landscape,
  JM,
  n.timestep,
  spp.freq,
  W.r,
  sigma,
  theta,
  root.value,
  scenario.ID,
  sim.ID,
  output.dir.path,
  reps = 1
)
}
\arguments{
\item{comm}{Matrix containing occurrences or abundances of species in sites. Species in columns and sites in rows.}

\item{subset}{Logical. If TRUE, only the subset of species present in comm will be considered in community simulations. Note that niche position will be simulated for all species in the phylogeny.}

\item{occurrence}{Logical. If TRUE, comm will be transformed to presence/absence matrix.}

\item{env}{Numeric vector containing values of the environmental variable for the sites.}

\item{site.coords}{Geographic coordinates (longitude and latitude) of sites.}

\item{tree}{Newick object containing a phylogenetic tree.}

\item{OU.alpha}{Numeric vector containing the values of alpha parameter of Ornstein-Uhlenbeck model.}

\item{sigma}{Numeric value indicating the sigma parameter to be used in \code{\link[ape:rTraitCont]{rTraitCont}} function.}

\item{theta}{Numeric value indicating the theta parameter to be used in \code{\link{rTraitCont}}.}

\item{root.value}{Numeric value indicating the root.value parameter to be used in \code{\link{rTraitCont}}.}

\item{runs}{Numeric value indicating the number of simulations to be performed.}

\item{ncores}{Numeric value indicating the number of cores to perform parallel computation. Default ncores = NULL.}

\item{area.m2}{A scalar or a vector. Area of each site to be used in metacommunity simulation. This only matters if immigration is defined as no. individuals / m2. Default value = 1.}

\item{m}{Immigration rate at each site reported as Hubbell’s m. Same as used in \code{\link{make.landscape}} of MCSim package. Default value = 0.5.}

\item{JM}{The number of individuals (abundance data) or species (presence/absence data) in the metacommunity. Same argument used in \code{\link{make.landscape}}.}

\item{JM.limit}{The maximum number of individuals in the metacommunity. Same argument used in \code{\link{make.landscape}}.}

\item{JL}{The number of individuals (abundance data) or species (presence/absence data) in each individual community. Same argument used in \code{\link{make.landscape}}.}

\item{nu}{The probability that a novel species will appear during a recruitment event. Same argument used in \code{\link{metasim}}. Default value = 0.}

\item{speciation.limit}{The speciation rate of simulated metacommunities. Same argument used in \code{\link{metasim}}. Default value = 0.}

\item{n.timestep}{Number of time steps (generations) in each simulation. Default value = 50.}

\item{W.r}{The slope of the dispersal kernel, indicating the intensity of dispersal of individuals among sites. Default = 0, indicating a panmitic metacommunity.}

\item{scenario.ID}{Provides the name of scenario that will be simulated.}

\item{sim.ID}{Provides a name for the simulation, which is saved along with parameter values in the output.dir.path.}

\item{output.dir.path}{Name of the folder that will be created in working directory of R session to store parameter values of simulations.}

\item{OU.alpha.v}{Numeric vector containing values of OU alpha parameter.}

\item{mat.env}{Matrix containing values for environmental gradient.}

\item{landscape}{Landscape where the simulations will go on. Object generated with \code{\link{make.landscape}} from MCSim package.}

\item{spp.freq}{Numeric vector of regional abundance, used in \code{\link{metasim}} from MCSim package.}

\item{reps}{Tricky argument to work in parallel process.}
}
\value{
List, with the following components:
\item{Entropy}{Entropy for empirical communities.}
\item{Predicted.entropy.1}{Mean predicted entropy of order 1 for simulated metacommunities.}
\item{Predicted.entropy.2}{Mean predicted entropy of order 2 for simulated metacommunities.}
\item{Predicted.entropy.12}{Mean predicted entropy of order 12 for simulated metacommunities.}
\item{K.niche.position}{Mean values for K statistic calculated with bootstrap procedure for simulated niche position.}
\item{Alpha.niche.position}{Mean values for alfa parameter of OU model calculated using bootstrap procedure for simulated niche position.}
\item{Z0.niche.position}{Mean values of simulated niche position at the root of the phylogeny calculated using bootstrap procedure.}
\item{AIC}{Numerical matrix containing AIC statistic for each OU alpha value and entropy order.}
\item{W}{Numerical matrix containing Akaike weights (wi) from model selection, for each OU alpha value and entropy order.}
\item{R2}{Numerical matrix containing R2 statistic from linear model relating the observed entropies to mean predicted entropy values, for each OU alpha value and entropy order.}
}
\description{
Mcfly function to estimate the influence of stabilizing niche selection on species diversity across environmental gradients
}
