% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcfly_function.R
\name{Mcfly}
\alias{Mcfly}
\title{Mcfly}
\usage{
Mcfly(
  comm,
  subset,
  occurrence = FALSE,
  env,
  site.coords,
  tree,
  OU.alpha,
  sigma,
  theta,
  root.value,
  runs,
  ncores,
  area.m2 = 1,
  m = 0.5,
  JM = sum(comm),
  JM.limit = JM,
  JL = rowSums(comm),
  nu = 0,
  speciation.limit = 0,
  n.timestep = 50,
  W.r = 0,
  scenario.ID = "species.sorting",
  sim.ID = "data",
  output.dir.path = "OUTPUT_DATA"
)
}
\arguments{
\item{comm}{Matrix containing occurence or abundance of species in samples. Species in columns and samples in rows}

\item{subset}{The subset of species presented in phylogenetic tree and in comm argument}

\item{occurrence}{Logical, if TRUE comm will be transformed to presence/absence matrix}

\item{env}{Numeric vector containing values for environmental variable of sites}

\item{site.coords}{X and Y coordinates for sites of community matrix}

\item{tree}{Newick object containing a phylogenetic tree}

\item{OU.alpha}{Numeric vector containing the values of alpha parameter of Ornstein Uhlenbeck model}

\item{sigma}{Numeric value indicating the sigma parameter to be used in \code{\link[ape:rTraitCont]{rTraitCont}} function}

\item{theta}{Numeric value indicating the theta parameter to be used in \code{\link{rTraitCont}}}

\item{root.value}{Numeric value indicating the root.value parameter to be used in \code{\link{rTraitCont}}}

\item{runs}{Numeric value indicating the number of simulations to be performed}

\item{ncores}{Numeric value indicating the number of cores to perform parallel computation. Generally the number of OU.alpha values}

\item{area.m2}{Area of a site to be used in metacommunity simulation. This only matters if immigration is defined as no. individuals / m2}

\item{m}{Immigration rate at each site reported as Hubbell’s m. Same as used in \code{\link{make.landscape}} of MCSim package}

\item{JM}{The number of individual organisms that inhabit the metacommunity. Same argument used in \code{\link{make.landscape}}}

\item{JM.limit}{The maximmun number of individual organisms that inhabit the metacommunity. Same argument used in \code{\link{make.landscape}}}

\item{JL}{The number of individuals that make each individual community. Same argument used in \code{\link{make.landscape}}}

\item{nu}{The probability that a novel species will appear during a recruitment event. Same argument used in \code{\link{metasim}}}

\item{speciation.limit}{The speciation rate of simulated metacommunities. Same argument used in \code{\link{metasim}}}

\item{n.timestep}{Number of generations in each simulation. Default is 50}

\item{W.r}{The slope of the dispersal kernel, indicating how individuals move among sites. Default is 0, indicating a panmitic metacommunity}

\item{scenario.ID}{Provides the name of scenario that will be simulated}

\item{sim.ID}{Provides a name for the simulation, which is saved along with parameter values in the output.dir.path}

\item{output.dir.path}{Folder that will be created in working directory of R session to store parameter values of simulations}
}
\value{
List, with the following components
\item{Entropy}{Entropy for empirical community}
\item{Predicted.entropy.1}{Mean predicted entropy with exponent 1 for simulated metacommunities}
\item{Predicted.entropy.2}{Mean predicted entropy with exponent 2 for simulated metacommunities}
\item{Predicted.entropy.12}{Mean predicted entropy with exponent 2 for simulated metacommunities}
\item{K.niche.position}{Mean values for K statistic calculated with bootstrap procedure from simulated metacommunities}
\item{Alpha.niche.position}{Mean values for alfa parameter of OU model calculated using bootstrap procedure from simulated metacommunities}
\item{Z0.niche.position}{Mean values for root value generated from of OU model calculated using bootstrap procedure from simulated metacommunities}
\item{AIC}{Numerical matrix containing AIC statistic for each alpha value, for each entropy}
\item{W}{Numerical matrix containing w statistic of model seelction for each alpha value, for each entropy}
\item{R2}{Numerical matrix containing R2 statistic from linear model relating the observed entropies with mean predicted values of entropies}
}
\description{
Mcfly function to estimate the influence of stabilizing niche selection on species diversity across environmental gradients
}
